refactoring: setter, builder -> getWalkInfo
